requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
backports.zoneinfo==0.2.1   
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0   
cryptography==41.0.1        
defusedxml==0.7.1
dj-database-url==2.0.0
Django==4.2.3
django-cors-headers==4.1.0
django-environ==0.10.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
environ==1.0
geographiclib==2.0
geopy==2.3.0
googlemaps==4.10.0
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
Pillow==9.1.1
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
whitenoise==6.5.0